Verified force vector with linear dload
dload function can be tricky, it needs to be defined along the frame, not along any projection.
If the function is defined along a projected line, then the actual distribution along the frame must firstly be derived based on conservation of total force vector.
If it takes in global coords, it outputs global tract vector. If it takes in local coords, it outputs local tract vector.