Added verify material, removed truss-material support in frame2d2elem
Added a verify material method for each element such that the material type can be checked before assembly the global K matrix.
The reason of removing truss-material support in frame2d2elem is because the dload force vector cannot be consistent unless it takes in material as argument. This is ugly and destroys all other elements' dload force vector functions. The right approach is to construct a more powerful dof_map support to allow elements with different dofs to be connected consistently. For that a global dof_list for a node is needed which includes all the possible dofs a node can have and their respective index in the global dof_list. Each element will have its own dof_list with indices such as [0,1] for truss2d, [0,1,5] for frame2d and [0,1,3,5] for frame2d with torsional.